Uppgifter/Examinering-1-ETL/requirements.txt
matplotlib==3.6.3
pandas==1.5.2
psycopg2==2.9.6
python-dotenv==1.0.0
requests==2.28.2
